+++ app/src/main/java/com/amanichristian/agl/AndroidOSms/Others/PartnerStatistics.java
@@ -1,6 +1,4 @@
-package com.amanichristian.agl.osmsandroid.Others;
-
-import com.amanichristian.agl.osmsandroid.Others.ServiceStatistics;
+package com.amanichristian.agl.AndroidOSms.Others;
 
 /**
  * Created by AMANI CHRISTIAN CYRILLE on 23/11/2015.
