+++ app/src/main/java/com/amanichristian/agl/AndroidOSms/Others/PurchaseOrders.java
@@ -1,4 +1,4 @@
-package com.amanichristian.agl.osmsandroid.Others;
+package com.amanichristian.agl.AndroidOSms.Others;
 
 /**
  * Created by AMANI CHRISTIAN CYRILLE on 23/11/2015.
