+++ app/src/main/java/com/amanichristian/agl/AndroidOSms/Others/ServiceStatistics.java
@@ -1,4 +1,4 @@
-package com.amanichristian.agl.osmsandroid.Others;
+package com.amanichristian.agl.AndroidOSms.Others;
 
 /**
  * Created by AMANI CHRISTIAN CYRILLE on 27/11/2015.
