+++ app/src/main/java/com/amanichristian/agl/AndroidOSms/RemainderSMS.java
@@ -1,6 +1,6 @@
-package com.amanichristian.agl.osmsandroid;
+package com.amanichristian.agl.AndroidOSms;
 
-import com.amanichristian.agl.osmsandroid.Others.PartnerContracts;
+import com.amanichristian.agl.AndroidOSms.Others.PartnerContracts;
 
 /**
  * Created by AMANI CHRISTIAN CYRILLE on 25/11/2015.
