+++ app/src/main/java/com/amanichristian/agl/AndroidOSms/ResponseSMS.java
@@ -1,4 +1,4 @@
-package com.amanichristian.agl.osmsandroid;
+package com.amanichristian.agl.AndroidOSms;
 
 /**
  * Created by AMANI CHRISTIAN CYRILLE on 25/11/2015.
