+++ app/src/main/java/com/amanichristian/agl/AndroidOSms/ServiceOSms.java
@@ -1,4 +1,4 @@
-package com.amanichristian.agl.osmsandroid;
+package com.amanichristian.agl.AndroidOSms;
 
 /**
  * Created by AMANI CHRISTIAN CYRILLE on 23/11/2015.
